Consolidate demo runtime crate.
